LF-1: Change price to float

+++ src/Entity/OrderInfo.php
@@ -24,7 +24,7 @@ class OrderInfo
     private $user;
 
     /**
-     * @ORM\Column(type="integer", nullable=true)
+     * @ORM\Column(type="float", nullable=true)
      */
     private $totalPrice;
 
@@ -116,12 +116,12 @@ class OrderInfo
         return $this;
     }
 
-    public function getTotalPrice(): ?int
+    public function getTotalPrice(): ?float
     {
         return $this->totalPrice;
     }
 
-    public function setTotalPrice(?int $totalPrice): self
+    public function setTotalPrice(?float $totalPrice): self
     {
         $this->totalPrice = $totalPrice;
